% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot.R
\name{dbplot_boxplot}
\alias{dbplot_boxplot}
\title{Boxplot}
\usage{
dbplot_boxplot(data, x, var, coef = 1.5)
}
\arguments{
\item{data}{A table (tbl)}

\item{x}{A discrete variable in which to group the boxplots}

\item{var}{A continuous variable}

\item{coef}{Length of the whiskers as multiple of IQR. Defaults to 1.5}
}
\value{
A ggplot object displaying boxplots for the specified variable grouped by x.
}
\description{
Uses dplyr operations to aggregate data and then `ggplot2`
to create the boxplot. Because of this approach,
the calculations automatically run inside the database if `data` has
a database or sparklyr connection. The `class()` of such tables
in R are: tbl_sql, tbl_dbi, tbl_spark

Requires database support for percentile/quantile functions.
See \code{\link{db_compute_boxplot}} for supported database backends.
}
\examples{
\dontrun{
library(DBI)
library(dplyr)
library(ggplot2)
con <- dbConnect(duckdb::duckdb(), ":memory:")
db_mtcars <- copy_to(con, mtcars, "mtcars")

db_mtcars |>
  dbplot_boxplot(am, mpg)

dbDisconnect(con)
}

}
\seealso{
\code{\link{dbplot_bar}}, \code{\link{dbplot_line}} ,
 \code{\link{dbplot_raster}}, \code{\link{dbplot_histogram}}
}
