% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram.R
\name{dbplot_histogram}
\alias{dbplot_histogram}
\title{Histogram}
\usage{
dbplot_histogram(data, x, bins = 30, binwidth = NULL)
}
\arguments{
\item{data}{A table (tbl)}

\item{x}{A continuous variable}

\item{bins}{Number of bins. Defaults to 30.}

\item{binwidth}{Fixed width for each bin, in the same units as the data. Overrides bins when specified}
}
\value{
A ggplot object displaying a histogram of the specified variable.
}
\description{
Uses dplyr operations to aggregate data and then `ggplot2`
to create the histogram. Because of this approach,
the calculations automatically run inside the database if `data` has
a database or sparklyr connection. The `class()` of such tables
in R are: tbl_sql, tbl_dbi, tbl_spark
}
\examples{
\dontrun{
library(DBI)
library(dplyr)
library(ggplot2)
con <- dbConnect(duckdb::duckdb(), ":memory:")
db_mtcars <- copy_to(con, mtcars, "mtcars")

# A ggplot histogram with 30 bins
db_mtcars |>
  dbplot_histogram(mpg)

# A ggplot histogram with bins of size 10
db_mtcars |>
  dbplot_histogram(mpg, binwidth = 10)

dbDisconnect(con)
}
}
\seealso{
\code{\link{dbplot_bar}}, \code{\link{dbplot_line}} ,
 \code{\link{dbplot_raster}}
}
