% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggvis-raster.R
\name{layer_raster}
\alias{layer_raster}
\title{Raster plot}
\usage{
layer_raster(vis, ..., fill = NULL, res = 50)
}
\arguments{
\item{vis}{Visualization to modify}

\item{...}{Visual properties, passed on to props.}

\item{fill}{The aggregation formula. Defaults to count (n)}

\item{res}{The number of bins created by variable. The highest the number, the more records can be potentially imported from the sourd}
}
\description{
To visualize two continuous variables, we typically resort to a Scatter plot. However, 
this may not be practical when visualizing millions or billions of dots representing the 
intersections of the two variables. A Raster plot may be a better option, 
because it concentrates the intersections into squares that are easier to parse visually.

This function will work with all data frame types, including tbl_sql
}
\details{
There are two considerations when using a Raster plot with a database. Both considerations are related 
to the size of the results downloaded from the database:

- The number of bins requested: The higher the bins value is, the more data is downloaded from the database.

- How concentrated the data is: This refers to how many intersections return a value. The more intersections without a value, 
the less data is downloaded from the database.
}
\examples{

library(ggvis)

mtcars \%>\% 
  ggvis(~mpg, ~cyl) \%>\% 
  layer_raster()

# Also supports aggregate formulas for the fill argument
mtcars \%>\%
  ggvis(~mpg, ~cyl) \%>\% 
 layer_raster(fill = ~sum(wt))

}
