% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbbin.R
\name{db_bin}
\alias{db_bin}
\title{Bin formula}
\usage{
db_bin(var, bins = 30, binwidth = NULL)
}
\arguments{
\item{var}{Variable name or formula}

\item{bins}{Number of bins. Defaults to 30.}

\item{binwidth}{Fixed width for each bin, in the same units as the data. Overrides bins when specified}
}
\value{
An unevaluated expression (rlang quosure) that calculates bin membership
  for the specified variable. This expression is designed to be used within dplyr
  verbs using the `!!` operator.
}
\description{
Uses the rlang package to build the formula needed to create the bins of a numeric variable
in an unevaluated fashion.  This way, the formula can be then passed inside a dplyr
verb.
}
\examples{
\dontrun{
library(DBI)
library(dplyr)
con <- dbConnect(duckdb::duckdb(), ":memory:")
db_mtcars <- copy_to(con, mtcars, "mtcars")

# Important: Always name the field and
# prefix the function with `!!` (See Details)

# Uses the default 30 bins
db_mtcars |>
  group_by(x = !!db_bin(mpg)) |>
  count()

# Uses binwidth which overrides bins
db_mtcars |>
  group_by(x = !!db_bin(mpg, binwidth = 10)) |>
  count()

dbDisconnect(con)
}
}
