% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot.R
\name{dbplot_boxplot}
\alias{dbplot_boxplot}
\title{Boxplot}
\usage{
dbplot_boxplot(data, x, var, coef = 1.5)
}
\arguments{
\item{data}{A table (tbl)}

\item{x}{A discrete variable in which to group the boxplots}

\item{var}{A continuous variable}

\item{coef}{Length of the whiskers as multiple of IQR. Defaults to 1.5}
}
\description{
Uses very generic dplyr code to aggregate data and then `ggplot2` 
to create the boxplot  Because of this approach,
the calculations automatically run inside the database if `data` has
a database or sparklyr connection. The `class()` of such tables
in R are: tbl_sql, tbl_dbi, tbl_spark

It currently only works with Spark and Hive connections.
}
\seealso{
\code{\link{dbplot_bar}}, \code{\link{dbplot_line}} ,
 \code{\link{dbplot_raster}}, \code{\link{dbplot_histogram}}
}
